Split csv and tbl sources into separatemodules
