+++ tpchgen-cli/src/csv.rs
@@ -1,5 +1,4 @@
-//! Implementations of [`Source`] for various different TPCH tables
-
+//! Implementations of [`Source`] for generating data in TBL format
 use super::generate::Source;
 use std::io::Write;
 use tpchgen::csv::{
@@ -9,31 +8,6 @@ use tpchgen::generators::{
     CustomerGenerator, LineItemGenerator, NationGenerator, OrderGenerator, PartGenerator,
     PartSupplierGenerator, RegionGenerator, SupplierGenerator,
 };
-
-/// Define a Source that writes the table in TBL format
-macro_rules! define_tbl_source {
-    ($SOURCE_NAME:ident, $GENERATOR_TYPE:ty) => {
-        pub struct $SOURCE_NAME {
-            inner: $GENERATOR_TYPE,
-        }
-
-        impl $SOURCE_NAME {
-            pub fn new(inner: $GENERATOR_TYPE) -> Self {
-                Self { inner }
-            }
-        }
-
-        impl Source for $SOURCE_NAME {
-            fn create(self, mut buffer: Vec<u8>) -> Vec<u8> {
-                for item in self.inner.iter() {
-                    // The default Display impl writes TBL format
-                    writeln!(&mut buffer, "{item}").expect("writing to memory is infallible");
-                }
-                buffer
-            }
-        }
-    };
-}
 
 /// Define a Source that writes the table in CSV format
 macro_rules! define_csv_source {
@@ -61,16 +35,6 @@ macro_rules! define_csv_source {
         }
     };
 }
-
-// Define .tbl sources for all tables
-define_tbl_source!(NationTblSource, NationGenerator<'static>);
-define_tbl_source!(RegionTblSource, RegionGenerator<'static>);
-define_tbl_source!(PartTblSource, PartGenerator<'static>);
-define_tbl_source!(SupplierTblSource, SupplierGenerator<'static>);
-define_tbl_source!(PartSuppTblSource, PartSupplierGenerator<'static>);
-define_tbl_source!(CustomerTblSource, CustomerGenerator<'static>);
-define_tbl_source!(OrderTblSource, OrderGenerator<'static>);
-define_tbl_source!(LineItemTblSource, LineItemGenerator<'static>);
 
 // Define .csv sources for all tables
 define_csv_source!(NationCsvSource, NationGenerator<'static>, NationCsv);
